# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handlers.AddressResponse:
    properties:
      city:
        type: string
      country:
        type: string
      postal_code:
        type: string
      street:
        type: string
    type: object
...
  internal_handlers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  internal_handlers.OrganizationResponse:
    properties:
      address:
        $ref: '#/definitions/internal_handlers.AddressResponse'
      contact_email:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      domain:
        type: string
      id:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/internal_handlers.OrganizationSettingsResponse'
      slug:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  internal_handlers.OrganizationSettingsResponse:
    properties:
      default_due_days:
        type: integer
      default_language:
        type: string
      min_checkfix_grade:
        type: string
      notification_emails:
        items:
          type: string
        type: array
      notifications_enabled:
        type: boolean
      require_checkfix:
        type: boolean
    type: object
...
  internal_handlers.PaginatedTemplatesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handlers.TemplateResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handlers.TemplateResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      default_passing_score:
        type: integer
      description:
        type: string
      estimated_minutes:
        type: integer
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      published_at:
        type: string
      tags:
        items:
          type: string
        type: array
      topics:
        items:
          $ref: '#/definitions/internal_handlers.TemplateTopicResponse'
        type: array
      updated_at:
        type: string
      usage_count:
        type: integer
      version:
        type: string
    type: object
  internal_handlers.TemplateTopicResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      order:
        type: integer
    type: object
...
  internal_handlers.UpdateAddressRequest:
    properties:
      city:
        type: string
      country:
        type: string
      postal_code:
        type: string
      street:
        type: string
    type: object
...
  internal_handlers.UpdateOrganizationRequest:
    properties:
      address:
        $ref: '#/definitions/internal_handlers.UpdateAddressRequest'
      contact_email:
        type: string
      contact_phone:
        type: string
      domain:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/internal_handlers.UpdateSettingsRequest'
    type: object
...
  internal_handlers.UpdateSettingsRequest:
    properties:
      default_due_days:
        type: integer
      default_language:
        type: string
      min_checkfix_grade:
        type: string
      notification_emails:
        items:
          type: string
        type: array
      notifications_enabled:
        type: boolean
      require_checkfix:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nisfix.io
    name: NisFix Support
  description: B2B Supplier Security Portal API - Companies manage supplier security
    requirements through questionnaires and CheckFix reports
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NisFix Backend API
  version: "1.0"
paths:
...
  /organization:
    get:
      consumes:
      - application/json
      description: Gets the current user's organization details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.OrganizationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current organization
      tags:
      - Organization
    patch:
      consumes:
      - application/json
      description: Updates the current user's organization
      parameters:
      - description: Organization updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - Organization
...
  /templates:
    get:
      consumes:
      - application/json
      description: Lists all available system questionnaire templates
      parameters:
      - description: Filter by category (ISO27001, GDPR, NIS2)
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.TemplateResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List system templates
      tags:
      - Templates
  /templates/{id}:
    get:
      consumes:
      - application/json
      description: Gets details of a specific template
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.TemplateResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get template details
      tags:
      - Templates
  /templates/organization:
    get:
      consumes:
      - application/json
      description: Lists custom templates created by the company
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PaginatedTemplatesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organization templates
      tags:
      - Templates
  /templates/search:
    get:
      consumes:
      - application/json
      description: Searches templates by name or description
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.PaginatedTemplatesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search templates
      tags:
      - Templates
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"